+++ MultiprocessingSpider/utils.py
@@ -78,22 +78,12 @@ class UAGenerator:
         'Windows 95', 'Windows 98', 'Windows 98; Win 9x 4.90', 'Windows CE',
         'Windows NT 4.0', 'Windows NT 5.0', 'Windows NT 5.01',
         'Windows NT 5.1', 'Windows NT 5.2', 'Windows NT 6.0', 'Windows NT 6.1',
-        'Windows NT 6.2',
+        'Windows NT 6.2', 'Windows NT 10.0'
     )
 
     linux_processors = ('i686', 'x86_64')
 
     mac_processors = ('Intel', 'PPC', 'U; Intel', 'U; PPC')
-
-    android_versions = (
-        '1.0', '1.1', '1.5', '1.6', '2.0', '2.0.1', '2.1', '2.2', '2.2.1', '2.2.2', '2.2.3', '2.3', '2.3.1', '2.3.2',
-        '2.3.3', '2.3.4', '2.3.5', '2.3.6', '2.3.7', '3.0', '3.1', '3.2', '3.2.1', '3.2.2', '3.2.3', '3.2.4', '3.2.5',
-        '3.2.6', '4.0', '4.0.1', '4.0.2', '4.0.3', '4.0.4', '4.1', '4.1.1', '4.1.2', '4.2', '4.2.1', '4.2.2', '4.3',
-        '4.3.1', '4.4', '4.4.1', '4.4.2', '4.4.3', '4.4.4', '5.0', '5.0.1', '5.0.2', '5.1', '5.1.1', '6.0', '6.0.1',
-        '7.0', '7.1', '7.1.1', '7.1.2', '8.0.0', '8.1.0', '9',
-    )
-
-    apple_devices = ('iPhone', 'iPad')
 
     ios_versions = (
         '3.1.3', '4.2.1', '5.1.1', '6.1.6', '7.1.2', '9.3.5', '9.3.6', '10.3.3', '10.3.4', '12.4',
@@ -123,13 +113,7 @@ class UAGenerator:
                          randint(version_from, version_to),
                          randint(build_from, build_to),
                          saf),
-            tmplt.format('Linux; {0}'.format(self.android_platform_token()),
-                         saf,
-                         randint(version_from, version_to),
-                         randint(build_from, build_to),
-                         saf),
         )
-
         return 'Mozilla/5.0 ' + choice(platforms)
 
     def ie(self):
@@ -155,13 +139,4 @@ class UAGenerator:
             choice(self.mac_processors),
             randint(5, 12),
             randint(0, 9),
-        )
-
-    def android_platform_token(self):
-        return 'Android {0}'.format(choice(self.android_versions))
-
-    def ios_platform_token(self):
-        return '{0}; CPU {0} OS {1} like Mac OS X'.format(
-            choice(self.apple_devices),
-            choice(self.ios_versions).replace('.', '_'),
         )